Loop capture takes a LoopProfile
The window size, the encoded size, the frame clock, the quality and the
frame cap were five constants the recorder read directly, so a caller who
wanted a portrait master had to copy the recorder. They are now one value,
LoopProfile, that LoopCapturePlugin::new takes; ::default() is the
docs-loop profile the fleet already captured with.

The plugin publishes the profile during build, before its arming gate, so
force_capture_resolution reads it as an Option<Res<_>> without any Startup
ordering deciding the window size, and the smoke path renders the framing
the capture will. A profile whose output equals its window encodes with no
scale filter, rather than resampling a master through itself.

CAPTURE_RESOLUTION is now (u32, u32) so the fleet default and the profile
field share one type.

+++ crates/nova_autopilot/src/capture.rs
@@ -39,9 +39,12 @@ pub const CAPTURE_ENV: &str = "NOVA_CAPTURE";
 /// under. Unset (or empty), a relative path stays relative to the CWD.
 pub const CAPTURE_DIR_ENV: &str = "NOVA_CAPTURE_DIR";
 
-/// Default capture resolution: 1920x1080, the 16:9 the web figures and
-/// thumbnails use (thumbnails share this capture and are sized down in CSS).
-pub const CAPTURE_RESOLUTION: (f32, f32) = (1920.0, 1080.0);
+/// Default capture resolution in pixels: 1920x1080, the 16:9 the web figures
+/// and thumbnails use (thumbnails share this capture and are sized down in
+/// CSS). A run that wants another framing selects it through
+/// [`LoopProfile::window_resolution`](crate::loops::LoopProfile) rather than
+/// by editing this.
+pub const CAPTURE_RESOLUTION: (u32, u32) = (1920, 1080);
 
 /// Whether this run is on the capture path ([`CAPTURE_ENV`] is set).
 ///
